0_referencia3_remotion/pipeline/requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.7.0
